feat: привел к общему numeric(20,6) все цены в БД Fixes #36

+++ backend/shared/models/trade.py
@@ -13,6 +13,6 @@ class Trade(Base):
     asset_id: Mapped[int] = mapped_column(ForeignKey('assets.id'))
     direction: Mapped[str] = mapped_column(Text)  # buy/sell
     quantity: Mapped[float] = mapped_column(Numeric, nullable=False)
-    price: Mapped[float] = mapped_column(Numeric, nullable=False)
+    price: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
     trade_time: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
